Split off accel charts, make ui driver smaller.
